requirements.txt
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
isodate==0.7.2